- arduino-neopixel-color-bars.ino: fix newBar/getUnrelatedColor by initializing ColorBar.color to 0, and constraining the getUnrelatedColor for-barNo probe loop to bars that already received a color, and to exit the loop when all bars (up to current bar) have colors COLOR_DISTANCE_REQ away from colorCandidate.
In other words, getUnrelatedColor() was broken, and this repairs it so works as desired.  The strategy is to create color bars with colors that are contrasty relative to each other, by choosing a set of colors at least COLOR_DISTANCE_REQ percent of MAX_COLOR_DISTANCE away from any color bar's color.  That is Euclidean distance in RGB 3-space, and this sort of works: occasionally colors are chosen of the same hue, but are brighter or dimmer and thus surpass COLOR_DISTANCE_REQ and are accepted.  This might be addressed by a future patch.